+++ src/resdk/query.py
@@ -174,8 +174,9 @@ class ResolweQuery:
     def _add_filter(self, filter_):
         """Add filtering parameters."""
         for key, value in filter_.items():
-            # 'sample' is called 'entity' in the backend.
-            key = key.replace("sample", "entity")
+            # Replace 'sample' with 'entity' if necessary.
+            if self.resource.rename_sample_to_entity:
+                key = key.replace("sample", "entity")
             value = self._dehydrate_resources(value)
             if self._non_string_iterable(value):
                 value = ",".join(map(str, value))